+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Features
 * **guardian**: Add last_deployed and battery_percent_internal fields for Guardians
 * **core**: Broadcast SNS message regarding updated event
+* **legacy**: Removed amplitude endpoint (keep returning random data to maintain mobile apps)
 
 ## 1.0.30 (2021-12-30)
 
@@ -19,6 +20,7 @@
 
 ### Bug Fixes
 * **guardian:** User can register and deploy a Guardian without the guardianCreator role
+
 
 ## 1.0.29 (2021-12-15)